Needs to be f32 instead

+++ src/data.rs
@@ -11,8 +11,8 @@ pub enum Data {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct ServerInitialization {
     pub server_name: String,
-    pub price_per_object: i64,
-    pub territory_lifetime: i64,
+    pub price_per_object: f32,
+    pub territory_lifetime: f32,
     pub territory_data: String,
     pub server_start_time: DateTime<Utc>
 }
